Workaround (stable on the ground)
Because of the small offset of a gyro, quadcopter becomes not stable during static standing on the ground. rs_ctrl service (using piid controller) tries to put a quadcopter to the desired position, according to gyro values. But in the ground, in static position, nothing is changing. That's why piid controller becomes unstable and provides too much torques to the motors, which can lead to a turnover of a quadcopter. 
Workaround for this: to put some noise thresholds to a controller.